Update model errors, add constraints, change base model name

+++ WeatherApp/WeatherApp/Models/Dto/Atmosphere.cs
@@ -3,7 +3,7 @@
     public class Atmosphere
     {
         public virtual double Temp { get; set; }
-        public virtual int Pressure { get; set; }
+        public virtual float Pressure { get; set; }
         public virtual int Humidity { get; set; }
         public virtual double TempMin { get; set; }
         public virtual double TempMax { get; set; }
